pypi: adding python pypi requirements.txt
This one has some real silliness since hey don't have a deterministic
field to check, but is just an open text field of URL relevant to the
project. There can be a million ways that projects use this, so I expect
this to need to be added to and refined over time.

Signed-off-by: Vincent Batts <vbatts@hashbangbash.com>